requirements.txt
docling
python-magic-bin # for windows
python-magic # for linux
llama-cpp-python
transformers